Crate esp_hal

no_std HAL implementations for the peripherals which are common among Espressif devices. Implements a number of the traits defined by [embedded-hal].

Feature Flags

debug — Enable debug features in the HAL (used for development).

log — Enable logging output using the log crate.

rt (enabled by default) — Enable runtime support.

vectored (enabled by default) — Enable interrupt vectoring.

Chip Support Feature Flags

esp32 — Target the ESP32.

esp32c2 — Target the ESP32-C2.

esp32c3 — Target the ESP32-C3.

esp32c6 — Target the ESP32-C6.

esp32h2 — Target the ESP32-H2.

esp32p4 — Target the ESP32-P4.

esp32s2 — Target the ESP32-S2.

esp32s3 — Target the ESP32-S3.

RISC-V Exclusive Feature Flags

direct-vectoring — Enable direct interrupt vectoring.

flip-link — Move the stack to start of RAM to get zero-cost stack overflow protection (ESP32-C6 and ESPS32-H2 only!).

interrupt-preemption — Enable interrupt preemption.

place-spi-driver-in-ram — Configuration for placing device drivers in the IRAM for faster access.

rv-init-data — Initialize the .data section of memory.

rv-zero-rtc-bss — Zero the .bss section of low-power memory.

rv-init-rtc-data — Initialize the .data section of low-power memory.

Trait Implementation Feature Flags

async — Enable support for asynchronous operation, with interfaces provided by embedded-hal-async and embedded-io-async.

defmt — Implement defmt::Format on certain types.

eh1 — Implement the traits defined in the 1.0.0 releases of embedded-hal and embedded-hal-nb for the relevant peripherals.

embedded-io — Implement the traits defined in embedded-io for certain peripherals.

ufmt — Implement the ufmt_write::uWrite trait for certain peripherals.

Embassy Feature Flags

embassy — Enable support for embassy, a modern asynchronous embedded framework.

embassy-executor-interrupt — Use the interrupt-mode embassy executor.

embassy-executor-thread — Use the thread-mode embassy executor.

embassy-integrated-timers — Uses hardware timers as alarms for the executors. Using this feature limits the number of executors to the number of hardware alarms provided by the time driver.

embassy-time-systick — Enable the embassy time driver using the SYSTIMER peripheral. The SYSTIMER peripheral has three alarams available for use.

embassy-time-timg0 — Enable the embassy time driver using the TIMG0 peripheral. The TIMG0 peripheral has two alarms available for use.

PSRAM Feature Flags

psram-2m — Use externally connected PSRAM (2MB).

psram-4m — Use externally connected PSRAM (4MB).

psram-8m — Use externally connected PSRAM (8MB).

psram-80mhz — PSRAM 80Mhz frequency support

Octal RAM Feature Flags

opsram-2m — Use externally connected Octal RAM (2MB).

opsram-4m — Use externally connected Octal RAM (4MB).

opsram-8m — Use externally connected Octal RAM (8MB).

opsram-16m — Use externally connected Octal RAM (16MB).

Re-exports

pub use self::gpio::IO;

pub use esp_riscv_rt;

pub use procmacros as macros;

pub use riscv;

Modules

clock

Clock Control

efuse

Reading of eFuses (ESP32-P4)

gpio

General Purpose I/Os

interrupt

Interrupt support

peripheral

Exclusive peripheral access

peripherals

Peripheral Instances

prelude

The prelude

rom

ESP ROM libraries

system

System Control

trapframe

State of the CPU saved when entering exception or interrupt

Structs

FlashSafeDma

FlashSafeDma

Enums

Cpu

Available CPU cores

Functions

get_core

Which core the application is currently executing on

Attribute Macros

entry

Attribute to declare the entry point of the program