Module esp_hal::analog::adc

Analog to Digital Converter (ADC)

The Analog to Digital Converter (ADC) is integrated on the chip, and is capable of measuring analog signals from specific analog I/O pins. One or more ADC units are available, depending on the device being used.

Example

let mut adc1_config = AdcConfig::new();
let mut adc1 = ADC::<ADC1>::new(peripherals.ADC1, adc1_config);
let mut pin = adc1_config.enable_pin(io.pins.gpio2.into_analog(), Attenuation::Attenuation11dB);

let mut delay = Delay::new(&clocks);

loop {
    let pin_value: u16 = nb::block!(adc1.read(&mut pin)).unwrap();
    println!("PIN2 ADC reading = {}", pin_value);

    delay.delay_ms(1500u32);
}

Structs

ADC

Analog-to-Digital Converter peripheral driver.

AdcConfig

Configuration for the ADC.

AdcPin

An I/O pin which can be read using the ADC.

Enums

AdcCalSource

Calibration source of the ADC.

Attenuation

The attenuation of the ADC pin.

Resolution

The sampling/readout resolution of the ADC.

Traits

AdcCalScheme

A trait abstracting over calibration methods.

AdcChannel

A helper trait to get the ADC channel of a compatible GPIO pin.