Module esp_hal::ecc

ECC Accelerator

Overview

Elliptic Curve Cryptography (ECC) is an approach to public-key cryptography based on the algebraic structure of elliptic curves. ECC allows smaller keys compared to RSA cryptography while providing equivalent security.

ECC Accelerator can complete various calculation based on different elliptic curves, thus accelerating ECC algorithm and ECC-derived algorithms (such as ECDSA).

Main features

ECC Accelerator supports:

Two different elliptic curves, namely P-192 and P-256 defined in FIPS 186-3.

Seven working modes.

Interrupt upon completion of calculation.

Availability on ESP32 family

The accelerator is available on ESP32-C2 and ESP32-C6.

Data representation

Inputs of the ECC hardware accelerator must be provided in big-endian representation. The driver handles the inner representation of the blocks.

Structs

Ecc

Enums

EllipticCurve

Error

ECC interface error

WorkMode