Struct esp_hal::efuse::Efuse

pub struct Efuse;

Implementations

impl Efuse

pub fn read_base_mac_address() -> [u8; 6]

Reads chip’s MAC address from the eFuse storage.

pub fn get_flash_encryption() -> bool

Get status of SPI boot encryption.

pub fn get_rwdt_multiplier() -> u8

Get the multiplier for the timeout value of the RWDT STAGE 0 register.

impl Efuse

pub fn read_field_le<T: Sized + 'static>(field: EfuseField) -> T

Read field value in a little-endian order

pub fn read_field_be<T: Sized + 'static>(field: EfuseField) -> T

Read field value in a big-endian order

impl Efuse

pub fn set_mac_address(mac: [u8; 6]) -> Result<(), SetMacError>

Set the base mac address

The new value will be returned by read_mac_address instead of the one hard-coded in eFuse. This does not persist across device resets.

Can only be called once. Returns Err([SetMacError::AlreadySet]) otherwise.

pub fn get_mac_address() -> [u8; 6]

Get base mac address

By default this reads the base mac address from eFuse, but it can be overriden by set_mac_address.