Struct esp_hal::etm::EtmChannel

#[non_exhaustive]
pub struct EtmChannel<const C: u8> {}

Unconfigured EtmChannel.

Implementations

impl EtmChannel<0>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 0>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<1>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 1>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<2>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 2>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<3>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 3>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<4>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 4>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<5>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 5>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<6>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 6>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<7>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 7>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<8>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 8>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<9>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 9>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<10>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 10>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<11>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 11>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<12>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 12>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<13>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 13>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<14>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 14>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<15>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 15>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<16>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 16>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<17>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 17>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<18>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 18>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<19>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 19>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<20>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 20>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<21>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 21>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<22>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 22>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<23>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 23>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<24>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 24>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<25>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 25>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<26>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 26>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<27>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 27>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<28>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 28>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<29>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 29>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<30>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 30>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<31>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 31>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<32>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 32>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<33>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 33>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<34>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 34>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<35>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 35>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<36>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 36>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<37>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 37>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<38>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 38>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<39>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 39>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<40>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 40>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<41>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 41>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<42>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 42>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<43>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 43>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<44>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 44>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<45>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 45>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<46>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 46>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<47>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 47>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<48>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 48>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.

impl EtmChannel<49>

pub fn setup<'a, E, T>( self, event: &'a E, task: &'a T ) -> EtmConfiguredChannel<'a, E, T, 49>
where E: EtmEvent, T: EtmTask,

Setup the channel

Enabled the channel and configures the assigned event and task.