Module esp_hal::gpio::etm

Event Task Matrix Function

Overview

GPIO supports ETM function, that is, the ETM task of GPIO can be triggered by the ETM event of any peripheral, or the ETM task of any peripheral can be triggered by the ETM event of GPIO.

The GPIO ETM provides eight task channels. The ETM tasks that each task channel can receive are:

SET: GPIO goes high when triggered

CLEAR: GPIO goes low when triggered

TOGGLE: GPIO toggle level when triggered.

GPIO has eight event channels, and the ETM events that each event channel can generate are:

RISE_EDGE: Indicates that the output signal of the corresponding GPIO has a rising edge

FALL_EDGE: Indicates that the output signal of the corresponding GPIO has a falling edge

ANY_EDGE: Indicates that the output signal of the corresponding GPIO is reversed

Example

let led_task = gpio_ext.channel0_task.toggle(&mut led);
let button_event = gpio_ext.channel0_event.falling_edge(button);

Structs

GpioEtmChannels

All the GPIO ETM channels

GpioEtmEventChannel

An ETM controlled GPIO event

GpioEtmEventChannelAny

Event for any edge

GpioEtmEventChannelFalling

Event for falling edge

GpioEtmEventChannelRising

Event for rising edge

GpioEtmTaskChannel

An ETM controlled GPIO task

GpioEtmTaskClear

Task for clear operation

GpioEtmTaskSet

Task for set operation

GpioEtmTaskToggle

Task for toggle operation