Struct esp_hal::i2s::I2sWriteDmaTransfer

pub struct I2sWriteDmaTransfer<'d, T, CH, BUFFER>
where
    T: RegisterAccess,
    CH: ChannelTypes,
{ /* private fields */ }

An in-progress DMA write transfer.

Implementations

impl<'d, T, CH, BUFFER> I2sWriteDmaTransfer<'d, T, CH, BUFFER>
where T: RegisterAccess, CH: ChannelTypes,

pub fn available(&mut self) -> usize

Amount of bytes which can be pushed. Only useful for circular DMA transfers

pub fn push(&mut self, data: &[u8]) -> Result<usize, Error>

Push bytes into the DMA buffer. Only useful for circular DMA transfers

pub fn push_with( &mut self, f: impl FnOnce(&mut [u8]) -> usize ) -> Result<usize, Error>

Push bytes into the DMA buffer via the given closure. The closure must return the actual number of bytes written. The closure might get called with a slice which is smaller than the total available buffer. Only useful for circular DMA transfers

pub fn stop( self ) -> Result<(BUFFER, I2sTx<'d, T, CH>), (DmaError, BUFFER, I2sTx<'d, T, CH>)>

Stop for the DMA transfer and return the buffer and the I2sTx instance.