Module esp_hal::ledc::timer

LEDC timer

Overview

The LEDC Timer module is a part of the LED Controller (LEDC) driver designed for ESP microcontrollers. It provides a high-level interface to configure and control individual timers of the LEDC peripheral.

The module allows precise and flexible control over timer configurations, duty cycles and frequencies, making it ideal for Pulse-Width Modulation (PWM) applications and LED lighting control.

Modules

config

Timer configuration

Structs

Timer

Timer struct

Enums

Error

Timer errors

LSClockSource

Clock source for LS Timers

Number

Timer number

Traits

TimerHW

Interface for HW configuration of timer

TimerIFace

Interface for Timers

TimerSpeed

Trait defining the type of timer source