Module esp_hal::mcpwm::operator

MCPWM operators

MCPWM peripheral - operator module

Overview

The operator module is part of the MCPWM peripheral driver for ESP chips. It is responsible for generating PWM (Pulse Width Modulation) signals and handling various aspects related to PWM signal generation.

This module provides flexibility in configuring the PWM outputs. Its implementation allows for motor control and other applications that demand accurate pulse timing and sophisticated modulation techniques.

Structs

DeadTimeCfg

Configuration for MCPWM Operator DeadTime It’s recommended to reference the technical manual for configuration

Operator

A MCPWM operator

PwmActions

Settings for what actions should be taken on timing events

PwmPin

A pin driven by an MCPWM operator

PwmPinConfig

Configuration describing how the operator generates a signal on a connected pin

PwmUpdateMethod

Settings for when PwmPin::set_timestamp takes effect

Enums

PWMStream

Input/Output Stream descriptor for each channel

UpdateAction

An action the operator applies to an output