Struct esp_hal::mcpwm::operator::Operator

pub struct Operator<const OP: u8, PWM> { /* private fields */ }

A MCPWM operator

The PWM Operator submodule has the following functions:

Generates a PWM signal pair, based on timing references obtained from the corresponding PWM timer.

Each signal out of the PWM signal pair includes a specific pattern of dead time. (Not yet implemented)

Superimposes a carrier on the PWM signal, if configured to do so. (Not yet implemented)

Handles response under fault conditions. (Not yet implemented)

Implementations

impl<const OP: u8, PWM: PwmPeripheral> Operator<OP, PWM>

pub fn set_timer<const TIM: u8>(&mut self, timer: &Timer<TIM, PWM>)

Select a Timer to be the timing reference for this operator

Note:

By default TIMER0 is used

pub fn set_deadtime(&mut self, cfg: &DeadTimeCfg)

Configures deadtime for this operator

pub fn with_pin_a<'d, Pin: OutputPin>( self, pin: impl Peripheral<P = Pin> + 'd, config: PwmPinConfig<true> ) -> PwmPin<'d, Pin, PWM, OP, true>

Use the A output with the given pin and configuration

pub fn with_pin_b<'d, Pin: OutputPin>( self, pin: impl Peripheral<P = Pin> + 'd, config: PwmPinConfig<false> ) -> PwmPin<'d, Pin, PWM, OP, false>

Use the B output with the given pin and configuration

pub fn with_pins<'d, PinA: OutputPin, PinB: OutputPin>( self, pin_a: impl Peripheral<P = PinA> + 'd, config_a: PwmPinConfig<true>, pin_b: impl Peripheral<P = PinB> + 'd, config_b: PwmPinConfig<false> ) -> (PwmPin<'d, PinA, PWM, OP, true>, PwmPin<'d, PinB, PWM, OP, false>)

Use both the A and the B output with the given pins and configurations