Struct esp_hal::mcpwm::operator::PwmActions

pub struct PwmActions<const IS_A: bool>(/* private fields */);

Settings for what actions should be taken on timing events

Note:

The hardware supports using a timestamp A event to trigger an action on output B or vice versa. For clearer ownership semantics this HAL does not support such configurations.

Implementations

impl<const IS_A: bool> PwmActions<IS_A>

pub const UP_ACTIVE_HIGH: Self = _

Using this setting together with a timer configured with PwmWorkingMode::Increase will set the output high for a duration proportional to the set timestamp.

pub const UP_DOWN_ACTIVE_HIGH: Self = _

Using this setting together with a timer configured with PwmWorkingMode::UpDown will set the output high for a duration proportional to the set timestamp.

pub const fn empty() -> Self

PwmActions with no UpdateActions set

pub const fn on_up_counting_timer_equals_zero( self, action: UpdateAction ) -> Self

Choose an UpdateAction for an UTEZ event

pub const fn on_up_counting_timer_equals_period( self, action: UpdateAction ) -> Self

Choose an UpdateAction for an UTEP event

pub const fn on_up_counting_timer_equals_timestamp( self, action: UpdateAction ) -> Self

Choose an UpdateAction for an UTEA/UTEB event

pub const fn on_up_counting_timer_equals_ch_timestamp<const CH_A: bool>( self, action: UpdateAction ) -> Self

Choose an UpdateAction for an UTEA/UTEB event where you can specify which of the A/B to use

pub const fn on_down_counting_timer_equals_zero( self, action: UpdateAction ) -> Self

Choose an UpdateAction for an DTEZ event

pub const fn on_down_counting_timer_equals_period( self, action: UpdateAction ) -> Self

Choose an UpdateAction for an DTEP event

pub const fn on_down_counting_timer_equals_timestamp( self, action: UpdateAction ) -> Self

Choose an UpdateAction for an DTEA/DTEB event

pub const fn on_down_counting_timer_equals_ch_timestamp<const CH_A: bool>( self, action: UpdateAction ) -> Self

Choose an UpdateAction for an DTEA/DTEB event where you can specify which of the A/B to use