Struct esp_hal::mcpwm::operator::PwmPin

pub struct PwmPin<'d, Pin, PWM, const OP: u8, const IS_A: bool> { /* private fields */ }

A pin driven by an MCPWM operator

Implementations

impl<'d, Pin: OutputPin, PWM: PwmPeripheral, const OP: u8, const IS_A: bool> PwmPin<'d, Pin, PWM, OP, IS_A>

pub fn update_fed(&self, cycles: u16)

Updates dead-time FED register

WARNING: FED is connected to the operator, and could be connected to another pin

pub fn update_red(&self, cycles: u16)

Updates dead-time RED register

WARNING: RED is connected to the operator, and could be connected to another pin

pub fn set_actions(&mut self, value: PwmActions<IS_A>)

Configure what actions should be taken on timing events

pub fn set_update_method(&mut self, update_method: PwmUpdateMethod)

Set how a new timestamp syncs with the timer

pub fn set_timestamp(&mut self, value: u16)

Write a new timestamp. The written value will take effect according to the set PwmUpdateMethod.

pub fn get_timestamp(&self) -> u16

Get the old timestamp. The value of the timestamp will take effect according to the set PwmUpdateMethod.