Struct esp_hal::mcpwm::timer::Timer

pub struct Timer<const TIM: u8, PWM> { /* private fields */ }

A MCPWM timer

Every timer of a particular MCPWM peripheral can be used as a timing reference for every Operator of that peripheral

Implementations

impl<const TIM: u8, PWM: PwmPeripheral> Timer<TIM, PWM>

pub fn start(&mut self, timer_config: TimerClockConfig<'_>)

Apply the given timer configuration.

Note:

The prescaler and period configuration will be applied immediately and before setting the PwmWorkingMode. If the timer is already running you might want to call Timer::stop and/or Timer::set_counter first (if the new period is larger than the current counter value this will cause weird behavior).

The hardware supports writing these settings in sync with certain timer events but this HAL does not expose these for now.

pub fn stop(&mut self)

Stop the timer in its current state

pub fn set_counter(&mut self, phase: u16, direction: CounterDirection)

Set the timer counter to the provided value

pub fn status(&self) -> (u16, CounterDirection)

Read the counter value and counter direction of the timer