Module esp_hal::pcnt::channel

PCNT - channel configuration

Overview

The channel module is part of the PCNT peripheral driver for ESP chips.

It provides a convenient and efficient way to configure and use individual channels of the PCNT peripheral of pulse counting and signal edge detection on ESP chips.

Structs

Channel

Config

Pulse Counter configuration for a single channel

PcntSource

PcntPin can be always high, always low, or an actual pin

Enums

CtrlMode

PCNT channel action on control level

EdgeMode

PCNT channel action on signal edge

Number

Channel number