Module esp_hal::pcnt::unit

PCNT - Unit module

Overview

The unit module is a part of the PCNT peripheral driver for ESP chips. It offers functionalities for configuring and utilizing specific units of the PCNT peripheral.

Each unit is identified by a unit number, such as Unit0, Unit1, Unit2, and so on. This module provides methods to configure a unit with specific settings, like low and high limits, thresholds, and optional filtering. Users can easily configure these units based on their requirements.

Structs

Config

Unit configuration

Events

Unit

Enums

Error

Unit errors

Number

Unit number

ZeroMode

the current status of the counter.