Module esp_hal::peripherals

Peripheral Instances

This module creates singleton instances for each of the various peripherals, and re-exports them to allow users to access and use them in their applications.

Should be noted that that the module also re-exports the Interrupt enum from the PAC, allowing users to handle interrupts associated with these peripherals.

Structs

ADC1

AES

ASSIST_DEBUG

BT

DMA

DS

ECC

EFUSE

GPIO

GPIO_SD

HMAC

HP_APM

HP_SYS

I2C0

I2C1

I2S0

IEEE802154

INTERRUPT_CORE0

INTPRI

IO_MUX

LEDC

LPWR

LP_ANA

LP_AON

LP_APM

LP_PERI

LP_TIMER

LP_WDT

MCPWM0

MEM_MONITOR

MODEM_LPCON

MODEM_SYSCON

OTP_DEBUG

PARL_IO

PAU

PCNT

PMU

Peripherals

RMT

RNG

RSA

SHA

SOC_ETM

SPI0

SPI1

SPI2

SYSTEM

SYSTIMER

TEE

TIMG0

TIMG1

TRACE0

TWAI0

UART0

UART1

UHCI0

USB_DEVICE

Enums

Interrupt

Enumeration of all the interrupts.