Struct esp_hal::peripherals::ADC

pub struct ADC { /* private fields */ }

Implementations

impl ADC

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <ADC as Deref>::Target = {0x500de000 as *const <esp32p4::ADC as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <ADC as Deref>::Target

Return the pointer to the register block