Struct esp_hal::peripherals::AES

pub struct AES { /* private fields */ }

Implementations

impl AES

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <AES as Deref>::Target = {0x60088000 as *const <esp32c6::AES as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <AES as Deref>::Target

Return the pointer to the register block