Struct esp_hal::peripherals::AXI_DMA

pub struct AXI_DMA { /* private fields */ }

Implementations

impl AXI_DMA

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <AXI_DMA as Deref>::Target = {0x5008a000 as *const <esp32p4::AXI_DMA as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <AXI_DMA as Deref>::Target

Return the pointer to the register block