Struct esp_hal::peripherals::DS

pub struct DS { /* private fields */ }

Implementations

impl DS

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <DS as Deref>::Target = {0x6008c000 as *const <esp32h2::DS as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <DS as Deref>::Target

Return the pointer to the register block