Struct esp_hal::peripherals::ECDSA

pub struct ECDSA { /* private fields */ }

Implementations

impl ECDSA

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <ECDSA as Deref>::Target = {0x50096000 as *const <esp32p4::ECDSA as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <ECDSA as Deref>::Target

Return the pointer to the register block