Struct esp_hal::peripherals::EXTMEM

pub struct EXTMEM { /* private fields */ }

Implementations

impl EXTMEM

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <EXTMEM as Deref>::Target = {0x61800000 as *const <esp32s2::EXTMEM as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <EXTMEM as Deref>::Target

Return the pointer to the register block