Struct esp_hal::peripherals::GPIO

pub struct GPIO { /* private fields */ }

Implementations

impl GPIO

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <GPIO as Deref>::Target = {0x3ff44000 as *const <esp32::GPIO as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <GPIO as Deref>::Target

Return the pointer to the register block