Struct esp_hal::peripherals::H264

pub struct H264 { /* private fields */ }

Implementations

impl H264

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <H264 as Deref>::Target = {0x50084000 as *const <esp32p4::H264 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <H264 as Deref>::Target

Return the pointer to the register block