Struct esp_hal::peripherals::H264_DMA

pub struct H264_DMA { /* private fields */ }

Implementations

impl H264_DMA

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <H264_DMA as Deref>::Target = {0x500a7000 as *const <esp32p4::H264_DMA as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <H264_DMA as Deref>::Target

Return the pointer to the register block