Struct esp_hal::peripherals::HP_APM

pub struct HP_APM { /* private fields */ }

Implementations

impl HP_APM

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <HP_APM as Deref>::Target = {0x60099000 as *const <esp32c6::HP_APM as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <HP_APM as Deref>::Target

Return the pointer to the register block