Struct esp_hal::peripherals::HP_SYS_CLKRST

pub struct HP_SYS_CLKRST { /* private fields */ }

Implementations

impl HP_SYS_CLKRST

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <HP_SYS_CLKRST as Deref>::Target = {0x500e6000 as *const <esp32p4::HP_SYS_CLKRST as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <HP_SYS_CLKRST as Deref>::Target

Return the pointer to the register block