Struct esp_hal::peripherals::I2C0

pub struct I2C0 { /* private fields */ }

Implementations

impl I2C0

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <I2C0 as Deref>::Target = {0x3ff53000 as *const <esp32::I2C0 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <I2C0 as Deref>::Target

Return the pointer to the register block