Struct esp_hal::peripherals::I2C1

pub struct I2C1 { /* private fields */ }

Implementations

impl I2C1

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <I2C1 as Deref>::Target = {0x500c5000 as *const <esp32p4::I2C1 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <I2C1 as Deref>::Target

Return the pointer to the register block