Struct esp_hal::peripherals::I3C_MST

pub struct I3C_MST { /* private fields */ }

Implementations

impl I3C_MST

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <I3C_MST as Deref>::Target = {0x500da000 as *const <esp32p4::I3C_MST as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <I3C_MST as Deref>::Target

Return the pointer to the register block