Struct esp_hal::peripherals::INTERRUPT_CORE1

pub struct INTERRUPT_CORE1 { /* private fields */ }

Implementations

impl INTERRUPT_CORE1

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <INTERRUPT_CORE1 as Deref>::Target = {0x600c2000 as *const <esp32s3::INTERRUPT_CORE1 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <INTERRUPT_CORE1 as Deref>::Target

Return the pointer to the register block