Struct esp_hal::peripherals::INTPRI

pub struct INTPRI { /* private fields */ }

Implementations

impl INTPRI

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <INTPRI as Deref>::Target = {0x600c5000 as *const <esp32c6::INTPRI as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <INTPRI as Deref>::Target

Return the pointer to the register block