Struct esp_hal::peripherals::IO_MUX

pub struct IO_MUX { /* private fields */ }

Implementations

impl IO_MUX

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <IO_MUX as Deref>::Target = {0x3f409000 as *const <esp32s2::IO_MUX as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <IO_MUX as Deref>::Target

Return the pointer to the register block