Struct esp_hal::peripherals::ISP

pub struct ISP { /* private fields */ }

Implementations

impl ISP

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <ISP as Deref>::Target = {0x500a1000 as *const <esp32p4::ISP as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <ISP as Deref>::Target

Return the pointer to the register block