Struct esp_hal::peripherals::JPEG

pub struct JPEG { /* private fields */ }

Implementations

impl JPEG

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <JPEG as Deref>::Target = {0x50086000 as *const <esp32p4::JPEG as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <JPEG as Deref>::Target

Return the pointer to the register block