Struct esp_hal::peripherals::LCD_CAM

pub struct LCD_CAM { /* private fields */ }

Implementations

impl LCD_CAM

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <LCD_CAM as Deref>::Target = {0x500dc000 as *const <esp32p4::LCD_CAM as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <LCD_CAM as Deref>::Target

Return the pointer to the register block