Struct esp_hal::peripherals::LP_ADC

pub struct LP_ADC { /* private fields */ }

Implementations

impl LP_ADC

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <LP_ADC as Deref>::Target = {0x50127000 as *const <esp32p4::LP_ADC as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <LP_ADC as Deref>::Target

Return the pointer to the register block