Struct esp_hal::peripherals::LP_AON

pub struct LP_AON { /* private fields */ }

Implementations

impl LP_AON

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <LP_AON as Deref>::Target = {0x600b1000 as *const <esp32h2::LP_AON as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <LP_AON as Deref>::Target

Return the pointer to the register block