Struct esp_hal::peripherals::LP_AON_CLKRST

pub struct LP_AON_CLKRST { /* private fields */ }

Implementations

impl LP_AON_CLKRST

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <LP_AON_CLKRST as Deref>::Target = {0x50111000 as *const <esp32p4::LP_AON_CLKRST as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <LP_AON_CLKRST as Deref>::Target

Return the pointer to the register block