Struct esp_hal::peripherals::LP_APM0

pub struct LP_APM0 { /* private fields */ }

Implementations

impl LP_APM0

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <LP_APM0 as Deref>::Target = {0x60099800 as *const <esp32c6::LP_APM0 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <LP_APM0 as Deref>::Target

Return the pointer to the register block