Struct esp_hal::peripherals::LP_GPIO

pub struct LP_GPIO { /* private fields */ }

Implementations

impl LP_GPIO

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <LP_GPIO as Deref>::Target = {0x5012a000 as *const <esp32p4::LP_GPIO as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <LP_GPIO as Deref>::Target

Return the pointer to the register block