Struct esp_hal::peripherals::LP_I2S0

pub struct LP_I2S0 { /* private fields */ }

Implementations

impl LP_I2S0

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <LP_I2S0 as Deref>::Target = {0x50125000 as *const <esp32p4::LP_I2S0 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <LP_I2S0 as Deref>::Target

Return the pointer to the register block