Struct esp_hal::peripherals::LP_IO

pub struct LP_IO { /* private fields */ }

Implementations

impl LP_IO

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <LP_IO as Deref>::Target = {0x600b2000 as *const <esp32c6::LP_IO as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <LP_IO as Deref>::Target

Return the pointer to the register block