Struct esp_hal::peripherals::LP_IO_MUX

pub struct LP_IO_MUX { /* private fields */ }

Implementations

impl LP_IO_MUX

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <LP_IO_MUX as Deref>::Target = {0x5012b000 as *const <esp32p4::LP_IO_MUX as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <LP_IO_MUX as Deref>::Target

Return the pointer to the register block