Struct esp_hal::peripherals::LP_UART

pub struct LP_UART { /* private fields */ }

Implementations

impl LP_UART

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <LP_UART as Deref>::Target = {0x50121000 as *const <esp32p4::LP_UART as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <LP_UART as Deref>::Target

Return the pointer to the register block