Struct esp_hal::peripherals::MCPWM0

pub struct MCPWM0 { /* private fields */ }

Implementations

impl MCPWM0

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <MCPWM0 as Deref>::Target = {0x60014000 as *const <esp32h2::MCPWM0 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <MCPWM0 as Deref>::Target

Return the pointer to the register block