Struct esp_hal::peripherals::MCPWM1

pub struct MCPWM1 { /* private fields */ }

Implementations

impl MCPWM1

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <MCPWM1 as Deref>::Target = {0x500c1000 as *const <esp32p4::MCPWM1 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <MCPWM1 as Deref>::Target

Return the pointer to the register block