Struct esp_hal::peripherals::MIPI_CSI_HOST

pub struct MIPI_CSI_HOST { /* private fields */ }

Implementations

impl MIPI_CSI_HOST

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <MIPI_CSI_HOST as Deref>::Target = {0x5009f000 as *const <esp32p4::MIPI_CSI_HOST as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <MIPI_CSI_HOST as Deref>::Target

Return the pointer to the register block