Struct esp_hal::peripherals::MODEM_SYSCON

pub struct MODEM_SYSCON { /* private fields */ }

Implementations

impl MODEM_SYSCON

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <MODEM_SYSCON as Deref>::Target = {0x600a5400 as *const <esp32h2::MODEM_SYSCON as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <MODEM_SYSCON as Deref>::Target

Return the pointer to the register block