Struct esp_hal::peripherals::OTP_DEBUG

pub struct OTP_DEBUG { /* private fields */ }

Implementations

impl OTP_DEBUG

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <OTP_DEBUG as Deref>::Target = {0x600b3c00 as *const <esp32h2::OTP_DEBUG as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <OTP_DEBUG as Deref>::Target

Return the pointer to the register block