Struct esp_hal::peripherals::PARL_IO

pub struct PARL_IO { /* private fields */ }

Implementations

impl PARL_IO

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <PARL_IO as Deref>::Target = {0x500cf000 as *const <esp32p4::PARL_IO as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <PARL_IO as Deref>::Target

Return the pointer to the register block