Struct esp_hal::peripherals::PAU

pub struct PAU { /* private fields */ }

Implementations

impl PAU

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <PAU as Deref>::Target = {0x60093000 as *const <esp32c6::PAU as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <PAU as Deref>::Target

Return the pointer to the register block