Struct esp_hal::peripherals::PMS

pub struct PMS { /* private fields */ }

Implementations

impl PMS

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <PMS as Deref>::Target = {0x3f4c1000 as *const <esp32s2::PMS as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <PMS as Deref>::Target

Return the pointer to the register block