Struct esp_hal::peripherals::PMU

pub struct PMU { /* private fields */ }

Implementations

impl PMU

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <PMU as Deref>::Target = {0x50115000 as *const <esp32p4::PMU as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <PMU as Deref>::Target

Return the pointer to the register block