Struct esp_hal::peripherals::PVT

pub struct PVT { /* private fields */ }

Implementations

impl PVT

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <PVT as Deref>::Target = {0x5009e000 as *const <esp32p4::PVT as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <PVT as Deref>::Target

Return the pointer to the register block