Struct esp_hal::peripherals::Peripherals

pub struct Peripherals {

    pub ADC1: ADC1,
    pub ADC2: ADC2,
    pub AES: AES,
    pub APB_CTRL: APB_CTRL,
    pub BB: BB,
    pub BT: BT,
    pub DAC1: DAC1,
    pub DAC2: DAC2,
    pub DMA: DMA,
    pub EFUSE: EFUSE,
    pub FLASH_ENCRYPTION: FLASH_ENCRYPTION,
    pub FRC_TIMER: FRC_TIMER,
    pub GPIO: GPIO,
    pub GPIO_SD: GPIO_SD,
    pub HINF: HINF,
    pub I2C0: I2C0,
    pub I2C1: I2C1,
    pub I2S0: I2S0,
    pub I2S1: I2S1,
    pub IO_MUX: IO_MUX,
    pub LEDC: LEDC,
    pub MCPWM0: MCPWM0,
    pub MCPWM1: MCPWM1,
    pub NRX: NRX,
    pub PCNT: PCNT,
    pub PSRAM: PSRAM,
    pub RMT: RMT,
    pub RNG: RNG,
    pub RSA: RSA,
    pub LPWR: LPWR,
    pub RTC_IO: RTC_IO,
    pub RTC_I2C: RTC_I2C,
    pub SDHOST: SDHOST,
    pub SHA: SHA,
    pub SLC: SLC,
    pub SLCHOST: SLCHOST,
    pub SPI0: SPI0,
    pub SPI1: SPI1,
    pub SPI2: SPI2,
    pub SPI3: SPI3,
    pub SYSTEM: SYSTEM,
    pub TIMG0: TIMG0,
    pub TIMG1: TIMG1,
    pub TWAI0: TWAI0,
    pub UART0: UART0,
    pub UART1: UART1,
    pub UART2: UART2,
    pub UHCI0: UHCI0,
    pub UHCI1: UHCI1,
    pub WIFI: WIFI,

}

Fields
ADC1: ADC1ADC2: ADC2AES: AESAPB_CTRL: APB_CTRLBB: BBBT: BTDAC1: DAC1DAC2: DAC2DMA: DMAEFUSE: EFUSEFLASH_ENCRYPTION: FLASH_ENCRYPTIONFRC_TIMER: FRC_TIMERGPIO: GPIOGPIO_SD: GPIO_SDHINF: HINFI2C0: I2C0I2C1: I2C1I2S0: I2S0I2S1: I2S1IO_MUX: IO_MUXLEDC: LEDCMCPWM0: MCPWM0MCPWM1: MCPWM1NRX: NRXPCNT: PCNTPSRAM: PSRAMRMT: RMTRNG: RNGRSA: RSALPWR: LPWRRTC_IO: RTC_IORTC_I2C: RTC_I2CSDHOST: SDHOSTSHA: SHASLC: SLCSLCHOST: SLCHOSTSPI0: SPI0SPI1: SPI1SPI2: SPI2SPI3: SPI3SYSTEM: SYSTEMTIMG0: TIMG0TIMG1: TIMG1TWAI0: TWAI0UART0: UART0UART1: UART1UART2: UART2UHCI0: UHCI0UHCI1: UHCI1WIFI: WIFI
Implementations

impl Peripherals

pub fn take() -> Self

Returns all the peripherals once

impl Peripherals

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.