Struct esp_hal::peripherals::RSA

pub struct RSA { /* private fields */ }

Implementations

impl RSA

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <RSA as Deref>::Target = {0x50092000 as *const <esp32p4::RSA as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <RSA as Deref>::Target

Return the pointer to the register block