Struct esp_hal::peripherals::RTC_I2C

pub struct RTC_I2C { /* private fields */ }

Implementations

impl RTC_I2C

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <RTC_I2C as Deref>::Target = {0x60008c00 as *const <esp32s3::RTC_I2C as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <RTC_I2C as Deref>::Target

Return the pointer to the register block