Struct esp_hal::peripherals::RTC_IO

pub struct RTC_IO { /* private fields */ }

Implementations

impl RTC_IO

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <RTC_IO as Deref>::Target = {0x3ff48400 as *const <esp32::RTC_IO as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <RTC_IO as Deref>::Target

Return the pointer to the register block