Struct esp_hal::peripherals::SDHOST

pub struct SDHOST { /* private fields */ }

Implementations

impl SDHOST

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SDHOST as Deref>::Target = {0x50083000 as *const <esp32p4::SDHOST as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SDHOST as Deref>::Target

Return the pointer to the register block