Struct esp_hal::peripherals::SHA

pub struct SHA { /* private fields */ }

Implementations

impl SHA

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SHA as Deref>::Target = {0x3ff03000 as *const <esp32::SHA as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SHA as Deref>::Target

Return the pointer to the register block