Struct esp_hal::peripherals::SLC

pub struct SLC { /* private fields */ }

Implementations

impl SLC

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SLC as Deref>::Target = {0x3ff58000 as *const <esp32::SLC as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SLC as Deref>::Target

Return the pointer to the register block