Struct esp_hal::peripherals::SOC_ETM

pub struct SOC_ETM { /* private fields */ }

Implementations

impl SOC_ETM

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SOC_ETM as Deref>::Target = {0x500d5000 as *const <esp32p4::SOC_ETM as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SOC_ETM as Deref>::Target

Return the pointer to the register block