Struct esp_hal::peripherals::SPI0

pub struct SPI0 { /* private fields */ }

Implementations

impl SPI0

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SPI0 as Deref>::Target = {0x60003000 as *const <esp32s3::SPI0 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SPI0 as Deref>::Target

Return the pointer to the register block