Struct esp_hal::peripherals::SPI1

pub struct SPI1 { /* private fields */ }

Implementations

impl SPI1

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SPI1 as Deref>::Target = {0x60002000 as *const <esp32c3::SPI1 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SPI1 as Deref>::Target

Return the pointer to the register block