Struct esp_hal::peripherals::SPI2

pub struct SPI2 { /* private fields */ }

Implementations

impl SPI2

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SPI2 as Deref>::Target = {0x3ff64000 as *const <esp32::SPI2 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SPI2 as Deref>::Target

Return the pointer to the register block