Struct esp_hal::peripherals::SYSCON

pub struct SYSCON { /* private fields */ }

Implementations

impl SYSCON

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SYSCON as Deref>::Target = {0x3f426000 as *const <esp32s2::SYSCON as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SYSCON as Deref>::Target

Return the pointer to the register block