Struct esp_hal::peripherals::SYSTEM

pub struct SYSTEM { /* private fields */ }

Implementations

impl SYSTEM

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <DPORT as Deref>::Target = {0x3ff00000 as *const <esp32::DPORT as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <DPORT as Deref>::Target

Return the pointer to the register block