Struct esp_hal::peripherals::SYSTIMER

pub struct SYSTIMER { /* private fields */ }

Implementations

impl SYSTIMER

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SYSTIMER as Deref>::Target = {0x500e2000 as *const <esp32p4::SYSTIMER as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SYSTIMER as Deref>::Target

Return the pointer to the register block