Struct esp_hal::peripherals::TIMG1

pub struct TIMG1 { /* private fields */ }

Implementations

impl TIMG1

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <TIMG1 as Deref>::Target = {0x3ff60000 as *const <esp32::TIMG1 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <TIMG1 as Deref>::Target

Return the pointer to the register block