Struct esp_hal::peripherals::TWAI0

pub struct TWAI0 { /* private fields */ }

Implementations

impl TWAI0

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <TWAI0 as Deref>::Target = {0x6000c000 as *const <esp32h2::TWAI0 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <TWAI0 as Deref>::Target

Return the pointer to the register block