Struct esp_hal::peripherals::UART0

pub struct UART0 { /* private fields */ }

Implementations

impl UART0

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <UART0 as Deref>::Target = {0x60000000 as *const <esp32c3::UART0 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <UART0 as Deref>::Target

Return the pointer to the register block