Struct esp_hal::peripherals::USB0

pub struct USB0 { /* private fields */ }

Implementations

impl USB0

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <USB0 as Deref>::Target = {0x60080000 as *const <esp32s2::USB0 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <USB0 as Deref>::Target

Return the pointer to the register block