Struct esp_hal::peripherals::USB_DEVICE

pub struct USB_DEVICE { /* private fields */ }

Implementations

impl USB_DEVICE

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <USB_DEVICE as Deref>::Target = {0x6000f000 as *const <esp32c6::USB_DEVICE as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <USB_DEVICE as Deref>::Target

Return the pointer to the register block