Struct esp_hal::peripherals::WCL

pub struct WCL { /* private fields */ }

Implementations

impl WCL

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <WCL as Deref>::Target = {0x600d0000 as *const <esp32s3::WCL as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <WCL as Deref>::Target

Return the pointer to the register block