Module esp_hal::prelude

The prelude

Re-exports all traits required for interacting with the various peripheral drivers implemented in this crate.

Re-exports

pub use crate::clock::Clock as _esp_hal_clock_Clock;

pub use crate::dma::DmaTransfer as _esp_hal_dma_DmaTransfer;

pub use crate::dma::DmaTransferRxTx as _esp_hal_dma_DmaTransferRxTx;

pub use crate::gpio::InputPin as _esp_hal_gpio_InputPin;

pub use crate::gpio::OutputPin as _esp_hal_gpio_OutputPin;

pub use crate::gpio::Pin as _esp_hal_gpio_Pin;

pub use crate::i2c::Instance as _esp_hal_i2c_Instance;

pub use crate::ledc::channel::ChannelHW as _esp_hal_ledc_channel_ChannelHW;

pub use crate::ledc::channel::ChannelIFace as _esp_hal_ledc_channel_ChannelIFace;

pub use crate::ledc::timer::TimerHW as _esp_hal_ledc_timer_TimerHW;

pub use crate::ledc::timer::TimerIFace as _esp_hal_ledc_timer_TimerIFace;

pub use crate::system::SystemExt as _esp_hal_system_SystemExt;

pub use crate::timer::Instance as _esp_hal_timer_Instance;

pub use crate::timer::TimerGroupInstance as _esp_hal_timer_TimerGroupInstance;

pub use crate::uart::Instance as _esp_hal_uart_Instance;

pub use crate::uart::UartPins as _esp_hal_uart_UartPins;

pub use nb;

Macros

load_lp_code

Load code to be run on the LP/ULP core.

make_gpio_enum_dispatch_macro

Create an enum for erased GPIO pins, using the enum-dispatch pattern

Traits

_embedded_dma_ReadBuffer

Trait for buffers that can be given to DMA for reading.

_embedded_dma_ReadTarget

Trait for Deref targets used by the blanket DmaReadBuffer impl.

_embedded_dma_Word

Trait for DMA word types used by the blanket DMA buffer impls.

_embedded_dma_WriteBuffer

Trait for buffers that can be given to DMA for writing.

_embedded_dma_WriteTarget

Trait for DerefMut targets used by the blanket DmaWriteBuffer impl.

_embedded_hal_Capture

Input capture

_embedded_hal_Pwm

Pulse Width Modulation

_embedded_hal_PwmPin

A single PWM channel / pin

_embedded_hal_Qei

Quadrature encoder interface

_embedded_hal_adc_OneShot

ADCs that sample on single channels per request, and do so at the time of the request.

_embedded_hal_blocking_delay_DelayMs

Millisecond delay

_embedded_hal_blocking_delay_DelayUs

Microsecond delay

_embedded_hal_blocking_i2c_Read

Blocking read

_embedded_hal_blocking_i2c_Write

Blocking write

_embedded_hal_blocking_i2c_WriteRead

Blocking write + read

_embedded_hal_blocking_rng_Read

Blocking read

_embedded_hal_blocking_serial_Write

Write half of a serial interface (blocking variant)

_embedded_hal_blocking_spi_Transfer

Blocking transfer

_embedded_hal_blocking_spi_Write

Blocking write

_embedded_hal_digital_InputPinDeprecated

Single digital input pin

_embedded_hal_digital_OutputPinDeprecated

Single digital push-pull output pin

_embedded_hal_digital_ToggleableOutputPinDeprecated

Output pin that can be toggled

_embedded_hal_digital_v2_InputPin

Single digital input pin

_embedded_hal_digital_v2_OutputPin

Single digital push-pull output pin

_embedded_hal_digital_v2_StatefulOutputPin

Push-pull output pin that can read its output state

_embedded_hal_digital_v2_ToggleableOutputPin

Output pin that can be toggled

_embedded_hal_serial_Read

Read half of a serial interface

_embedded_hal_serial_Write

Write half of a serial interface

_embedded_hal_spi_FullDuplex

Full duplex (master mode)

_embedded_hal_timer_CountDown

A count down timer

_embedded_hal_watchdog_Watchdog

Feeds an existing watchdog to ensure the processor isn’t reset. Sometimes commonly referred to as “kicking” or “refreshing”.

_embedded_hal_watchdog_WatchdogDisable

Disables a running watchdog timer so the processor won’t be reset.

_embedded_hal_watchdog_WatchdogEnable

Enables A watchdog timer to reset the processor if software is frozen or stalled.

_fugit_ExtU32

Extension trait for simple short-hands for u32 Durations

_fugit_ExtU64

Extension trait for simple short-hands for u64 Durations

_fugit_RateExtU32

Extension trait for simple short-hands for u32 Rate

_fugit_RateExtU64

Extension trait for simple short-hands for u64 Rate

Attribute Macros

entry

Marks a function as the main function to be called on program start

interrupt

Marks a function as an interrupt handler

ram

This attribute allows placing statics and functions into ram.