Struct esp_hal::rmt::PulseCode

pub struct PulseCode {
    pub level1: bool,
    pub length1: u16,
    pub level2: bool,
    pub length2: u16,
}

Convenience representation of a pulse code entry.

Allows for the assignment of two levels and their lengths

Fields
level1: bool

Logical output level in the first pulse code interval

length1: u16

Length of the first pulse code interval (in clock cycles)

level2: bool

Logical output level in the second pulse code interval

length2: u16

Length of the second pulse code interval (in clock cycles)