Trait esp_hal::rmt::TxChannel

pub trait TxChannel: TxChannelInternal {
    // Provided methods
    fn transmit<T: Into<u32> + Copy>(
        self,
        data: &[T]
    ) -> SingleShotTxTransaction<'_, Self, T>
       where Self: Sized { ... }
    fn transmit_continuously<T: Into<u32> + Copy>(
        self,
        data: &[T]
    ) -> Result<ContinuousTxTransaction<Self>, Error>
       where Self: Sized { ... }
    fn transmit_continuously_with_loopcount<T: Into<u32> + Copy>(
        self,
        loopcount: u16,
        data: &[T]
    ) -> Result<ContinuousTxTransaction<Self>, Error>
       where Self: Sized { ... }
}

Provided Methods

fn transmit<T: Into<u32> + Copy>( self, data: &[T] ) -> SingleShotTxTransaction<'_, Self, T>
where Self: Sized,

Start transmitting the given pulse code sequence. This returns a SingleShotTxTransaction which can be used to wait for the transaction to complete and get back the channel for further use.

fn transmit_continuously<T: Into<u32> + Copy>( self, data: &[T] ) -> Result<ContinuousTxTransaction<Self>, Error>
where Self: Sized,

Start transmitting the given pulse code continuously. This returns a ContinuousTxTransaction which can be used to stop the ongoing transmission and get back the channel for further use. The length of sequence cannot exceed the size of the allocated RMT RAM.

fn transmit_continuously_with_loopcount<T: Into<u32> + Copy>( self, loopcount: u16, data: &[T] ) -> Result<ContinuousTxTransaction<Self>, Error>
where Self: Sized,

Like Self::transmit_continuously but also sets a loop count. ContinuousTxTransaction can be used to check if the loop count is reached.

Object Safety

This trait is not object safe.