Enum esp_hal::rtc_cntl::SocResetReason

pub enum SocResetReason {

    ChipPowerOn = 1,
    CoreSw = 3,
    CoreDeepSleep = 5,
    CoreMwdt0 = 7,
    CoreRtcWdt = 9,
    Cpu0Mwdt0 = 11,
    Cpu0Sw = 12,
    Cpu0RtcWdt = 13,
    SysBrownOut = 15,
    SysRtcWdt = 16,
    SysSuperWdt = 18,
    SysClkGlitch = 19,
    CoreEfuseCrc = 20,
    Cpu0Jtag = 24,

}

Variants

ChipPowerOn = 1

Power on reset

CoreSw = 3

Software resets the digital core by RTC_CNTL_SW_SYS_RST

CoreDeepSleep = 5

Deep sleep reset the digital core

CoreMwdt0 = 7

Main watch dog 0 resets digital core

CoreRtcWdt = 9

RTC watch dog resets digital core

Cpu0Mwdt0 = 11

Main watch dog 0 resets CPU 0

Cpu0Sw = 12

Software resets CPU 0 by RTC_CNTL_SW_PROCPU_RST

Cpu0RtcWdt = 13

RTC watch dog resets CPU 0

SysBrownOut = 15

VDD voltage is not stable and resets the digital core

SysRtcWdt = 16

RTC watch dog resets digital core and rtc module

SysSuperWdt = 18

Super watch dog resets the digital core and rtc module

SysClkGlitch = 19

Glitch on clock resets the digital core and rtc module

CoreEfuseCrc = 20

eFuse CRC error resets the digital core

Cpu0Jtag = 24

JTAG resets the CPU 0

Implementations

impl SocResetReason

pub const fn from_repr(discriminant: usize) -> Option<SocResetReason>

Try to create Self from the raw representation