Struct esp_hal::rtc_cntl::Rtc

pub struct Rtc<'d> {
    pub rwdt: Rwdt,
    pub swd: Swd,
    /* private fields */
}

Low-power Management

Fields
rwdt: Rwdtswd: Swd
Implementations

impl<'d> Rtc<'d>

pub fn new(rtc_cntl: impl Peripheral<P = LPWR> + 'd) -> Self

pub fn estimate_xtal_frequency(&mut self) -> u32

pub fn get_time_raw(&self) -> u64

read the current value of the rtc time registers.

pub fn get_time_us(&self) -> u64

read the current value of the rtc time registers in microseconds.

pub fn get_time_ms(&self) -> u64

read the current value of the rtc time registers in milliseconds

pub fn sleep_deep( &mut self, wake_sources: &[&dyn WakeSource], delay: &mut Delay ) -> !

enter deep sleep and wake with the provided wake_sources

pub fn sleep_light( &mut self, wake_sources: &[&dyn WakeSource], delay: &mut Delay )

enter light sleep and wake with the provided wake_sources

pub fn sleep( &mut self, config: &RtcSleepConfig, wake_sources: &[&dyn WakeSource], delay: &mut Delay )

enter sleep with the provided config and wake with the provided wake_sources