Module esp_hal::twai::filter

Two-wire Automotive Interface (TWAI) Filters

These are acceptance filters that limit which packets are received by the TWAI peripheral.

Structs

DualExtendedFilter

NOTE: The dual extended id acceptance filters can only match “the first 16 bits of the 29-bit ID”.

DualStandardFilter

A filter that matches against two standard 11-bit standard ids.

SingleExtendedFilter

Warning: This is not a perfect filter. Standard ids that match the bit layout of this filter will also be accepted.

SingleStandardFilter

A filter that matches against a single 11 bit id, the rtr bit, and the first two bytes of the payload.

Enums

FilterType

Traits

Filter

Type Aliases

BitFilter