Struct esp_hal::twai::Twai

pub struct Twai<'d, T> { /* private fields */ }

An active TWAI peripheral in Normal Mode.

In this mode, the TWAI controller can transmit and receive messages including error signals (such as error and overload frames).

Implementations

impl<'d, T> Twai<'d, T>
where T: Instance,

pub fn stop(self) -> TwaiConfiguration<'d, T>

Stop the peripheral, putting it into reset mode and enabling reconfiguration.

pub fn receive_error_count(&self) -> u8

pub fn transmit_error_count(&self) -> u8

pub fn is_bus_off(&self) -> bool

Check if the controller is in a bus off state.

pub fn num_available_messages(&self) -> u8

Get the number of messages that the peripheral has available in the receive FIFO.

Note that this may not be the number of valid messages in the receive FIFO due to fifo overflow/overrun.

pub fn clear_receive_fifo(&self)

Clear the receive FIFO, discarding any valid, partial, or invalid packets.

This is typically used to clear an overrun receive FIFO.

TODO: Not sure if this needs to be guarded against Bus Off or other error states.