Struct esp_hal::usb_serial_jtag::UsbSerialJtag

pub struct UsbSerialJtag<'d> { /* private fields */ }

USB Serial JTAG driver

Implementations

impl<'d> UsbSerialJtag<'d>

pub fn new(_usb_device: impl Peripheral<P = USB_DEVICE> + 'd) -> Self

Create a new USB serial/JTAG instance with defaults

pub fn split(self) -> (UsbSerialJtagTx<'d>, UsbSerialJtagRx<'d>)

Split the USB Serial JTAG peripheral into a transmitter and receiver, which is particuarly useful when having two tasks correlating to transmitting and receiving.

pub fn write_bytes(&mut self, data: &[u8]) -> Result<(), Infallible>

Write data to the serial output in chunks of up to 64 bytes

pub fn write_byte_nb(&mut self, word: u8) -> Result<(), Infallible>

Write data to the serial output in a non-blocking manner Requires manual flushing (automatically flushed every 64 bytes)

pub fn flush_tx(&mut self) -> Result<(), Infallible>

Flush the output FIFO and block until it has been sent

pub fn flush_tx_nb(&mut self) -> Result<(), Infallible>

Flush the output FIFO but don’t block if it isn’t ready immediately

pub fn read_byte(&mut self) -> Result<u8, Infallible>

pub fn listen_rx_packet_recv_interrupt(&mut self)

Listen for RX-PACKET-RECV interrupts

pub fn unlisten_rx_packet_recv_interrupt(&mut self)

Stop listening for RX-PACKET-RECV interrupts

pub fn rx_packet_recv_interrupt_set(&mut self) -> bool

Checks if RX-PACKET-RECV interrupt is set

pub fn reset_rx_packet_recv_interrupt(&mut self)

Reset RX-PACKET-RECV interrupt