Struct esp_hal::usb_serial_jtag::UsbSerialJtagRx

pub struct UsbSerialJtagRx<'d> { /* private fields */ }

USB Serial JTAG RX driver

Implementations

impl<'d> UsbSerialJtagRx<'d>

pub fn read_byte(&mut self) -> Result<u8, Infallible>

pub fn drain_rx_fifo(&mut self, buf: &mut [u8]) -> usize

Read all available bytes from the RX FIFO into the provided buffer and returns the number of read bytes. Never blocks. May stop early if the number of bytes in the FIFO is larger than buf.

pub fn listen_rx_packet_recv_interrupt(&mut self)

Listen for RX-PACKET-RECV interrupts

pub fn unlisten_rx_packet_recv_interrupt(&mut self)

Stop listening for RX-PACKET-RECV interrupts

pub fn rx_packet_recv_interrupt_set(&mut self) -> bool

Checks if RX-PACKET-RECV interrupt is set

pub fn reset_rx_packet_recv_interrupt(&mut self)

Reset RX-PACKET-RECV interrupt