Module esp_hal::efuse

Reading of eFuses (ESP32-P4)

Structs

Efuse

EfuseField

The bit field for get access to efuse data

Constants

BLOCK_SYS_DATA1

[SYS_DATA_PART1] System data part 1

CRYPT_DPA_ENABLE

[] Represents whether anti-dpa attack is enabled. 1:enabled. 0: disabled

DCDC_VSET

[] Set the dcdc voltage default

DCDC_VSET_EN

[] Select dcdc vset use efuse_dcdc_vset

DIS_DIRECT_BOOT

[] Represents whether direct boot mode is disabled or enabled. 1: disabled. 0: enabled

DIS_DOWNLOAD_MANUAL_ENCRYPT

[] Represents whether flash encrypt function is disabled or enabled(except in SPI boot mode). 1: disabled. 0: enabled

DIS_DOWNLOAD_MODE

[] Represents whether Download mode is disabled or enabled. 1: disabled. 0: enabled

DIS_FORCE_DOWNLOAD

[] Represents whether the function that forces chip into download mode is disabled or enabled. 1: disabled. 0: enabled

DIS_PAD_JTAG

[] Represents whether JTAG is disabled in the hard way(permanently). 1: disabled. 0: enabled

DIS_SWD

[] Set this bit to disable super-watchdog

DIS_TWAI

[] Represents whether TWAI function is disabled or enabled. 1: disabled. 0: enabled

DIS_USB_JTAG

[] Represents whether the function of usb switch to jtag is disabled or enabled. 1: disabled. 0: enabled

DIS_USB_OTG_DOWNLOAD_MODE

[] Set this bit to disable download via USB-OTG

DIS_USB_SERIAL_JTAG_DOWNLOAD_MODE

[] Represents whether the USB-Serial-JTAG download function is disabled or enabled. 1: disabled. 0: enabled

DIS_USB_SERIAL_JTAG_ROM_PRINT

[] Represents whether print from USB-Serial-JTAG is disabled or enabled. 1: disabled. 0: enabled

DIS_WDT

[] Set this bit to disable watch dog

ECDSA_ENABLE_SOFT_K

[] Represents whether hardware random number k is forced used in ESDCA. 1: force used. 0: not force used

ENABLE_SECURITY_DOWNLOAD

[] Represents whether security download is enabled or disabled. 1: enabled. 0: disabled

FLASH_ECC_EN

[] Set this bit to enable ecc for flash boot

FLASH_PAGE_SIZE

[] Set flash page size

FLASH_TPUW

[] Represents the flash waiting time after power-up; in unit of ms. When the value less than 15; the waiting time is the programmed value. Otherwise; the waiting time is 2 times the programmed value

FLASH_TYPE

[] The type of interfaced flash. 0: four data lines; 1: eight data lines

FORCE_DISABLE_SW_INIT_KEY

[] Set this bit to disable software written init key; and force use efuse_init_key

FORCE_SEND_RESUME

[] Represents whether ROM code is forced to send a resume command during SPI boot. 1: forced. 0:not forced

FORCE_USE_KEY_MANAGER_KEY

[] Set each bit to control whether corresponding key must come from key manager.. 1 is true; 0 is false. Bit0: ecdsa. Bit1: xts. Bit2: hmac. Bit3: ds

HP_PWR_SRC_SEL

[] HP system power source select. 0:LDO. 1: DCDC

HYS_EN_PAD

[] Represents whether the hysteresis function of corresponding PAD is enabled. 1: enabled. 0:disabled

JTAG_SEL_ENABLE

[] Represents whether the selection between usb_to_jtag and pad_to_jtag through strapping gpio15 when both EFUSE_DIS_PAD_JTAG and EFUSE_DIS_USB_JTAG are equal to 0 is enabled or disabled. 1: enabled. 0: disabled

KEY0

[BLOCK_KEY0] Key0 or user data

KEY1

[BLOCK_KEY1] Key1 or user data

KEY2

[BLOCK_KEY2] Key2 or user data

KEY3

[BLOCK_KEY3] Key3 or user data

KEY4

[BLOCK_KEY4] Key4 or user data

KEY5

[BLOCK_KEY5] Key5 or user data

KEY_PURPOSE_0

[KEY0_PURPOSE] Represents the purpose of Key0

KEY_PURPOSE_1

[KEY1_PURPOSE] Represents the purpose of Key1

KEY_PURPOSE_2

[KEY2_PURPOSE] Represents the purpose of Key2

KEY_PURPOSE_3

[KEY3_PURPOSE] Represents the purpose of Key3

KEY_PURPOSE_4

[KEY4_PURPOSE] Represents the purpose of Key4

KEY_PURPOSE_5

[KEY5_PURPOSE] Represents the purpose of Key5

KM_DEPLOY_ONLY_ONCE

[] Set each bit to control whether corresponding key can only be deployed once. 1 is true; 0 is false. Bit0: ecdsa. Bit1: xts. Bit2: hmac. Bit3: ds

KM_DISABLE_DEPLOY_MODE

[] TBD

KM_HUK_GEN_STATE_HIGH

[] Set this bit to control validation of HUK generate mode. Odd of 1 is invalid; even of 1 is valid

KM_HUK_GEN_STATE_LOW

[] Set this bit to control validation of HUK generate mode. Odd of 1 is invalid; even of 1 is valid

KM_RND_SWITCH_CYCLE

[] Set bits to control key manager random number switch cycle. 0: control by register. 1: 8 km clk cycles. 2: 16 km cycles. 3: 32 km cycles

LOCK_KM_KEY

[] TBD

MAC

[MAC_FACTORY] MAC address

MAC_EXT

[] Stores the extended bits of MAC address [0]

POWERGLITCH_EN

[] Represents whether power glitch function is enabled. 1: enabled. 0: disabled

PXA0_TIEH_SEL_0

[] TBD

PXA0_TIEH_SEL_1

[] TBD

PXA0_TIEH_SEL_2

[] TBD

PXA0_TIEH_SEL_3

[] TBD

RD_DIS

[] Disable reading from BlOCK4-10

RD_DIS_BLOCK_KEY0

[RD_DIS.KEY0] rd_dis of BLOCK_KEY0

RD_DIS_BLOCK_KEY1

[RD_DIS.KEY1] rd_dis of BLOCK_KEY1

RD_DIS_BLOCK_KEY2

[RD_DIS.KEY2] rd_dis of BLOCK_KEY2

RD_DIS_BLOCK_KEY3

[RD_DIS.KEY3] rd_dis of BLOCK_KEY3

RD_DIS_BLOCK_KEY4

[RD_DIS.KEY4] rd_dis of BLOCK_KEY4

RD_DIS_BLOCK_KEY5

[RD_DIS.KEY5] rd_dis of BLOCK_KEY5

RD_DIS_BLOCK_SYS_DATA2

[RD_DIS.SYS_DATA_PART2] rd_dis of BLOCK_SYS_DATA2

SECURE_BOOT_AGGRESSIVE_REVOKE

[] Represents whether revoking aggressive secure boot is enabled or disabled. 1: enabled. 0: disabled

SECURE_BOOT_DISABLE_FAST_WAKE

[] Represents whether FAST VERIFY ON WAKE is disabled or enabled when Secure Boot is enabled. 1: disabled. 0: enabled

SECURE_BOOT_EN

[] Represents whether secure boot is enabled or disabled. 1: enabled. 0: disabled

SECURE_BOOT_KEY_REVOKE0

[] Revoke 1st secure boot key

SECURE_BOOT_KEY_REVOKE1

[] Revoke 2nd secure boot key

SECURE_BOOT_KEY_REVOKE2

[] Revoke 3rd secure boot key

SECURE_VERSION

[] Represents the version used by ESP-IDF anti-rollback feature

SEC_DPA_LEVEL

[] Represents the spa secure level by configuring the clock random divide mode

SOFT_DIS_JTAG

[] Represents whether JTAG is disabled in soft way. Odd number: disabled. Even number: enabled

SPI_BOOT_CRYPT_CNT

[] Enables flash encryption when 1 or 3 bits are set and disables otherwise {0: “Disable”; 1: “Enable”; 3: “Disable”; 7: “Enable”}

SPI_DOWNLOAD_MSPI_DIS

[] Set this bit to disable accessing MSPI flash/MSPI ram by SYS AXI matrix during boot_mode_download

SYS_DATA_PART2

[BLOCK_SYS_DATA2] System data part 2 (reserved)

UART_PRINT_CONTROL

[] Represents the type of UART printing. 00: force enable printing. 01: enable printing when GPIO8 is reset at low level. 10: enable printing when GPIO8 is reset at high level. 11: force disable printing

USB_DEVICE_EXCHG_PINS

[] Enable usb device exchange pins of D+ and D-

USB_OTG11_EXCHG_PINS

[] Enable usb otg11 exchange pins of D+ and D-

USB_PHY_SEL

[] TBD

USER_DATA

[BLOCK_USR_DATA] User data

USER_DATA_MAC_CUSTOM

[MAC_CUSTOM CUSTOM_MAC] Custom MAC (TODO

WDT_DELAY_SEL

[] Represents whether RTC watchdog timeout threshold is selected at startup. 1: selected. 0: not selected

WR_DIS

[] Disable programming of individual eFuses

WR_DIS_BLK1

[] wr_dis of BLOCK1

WR_DIS_BLOCK_KEY0

[WR_DIS.KEY0] wr_dis of BLOCK_KEY0

WR_DIS_BLOCK_KEY1

[WR_DIS.KEY1] wr_dis of BLOCK_KEY1

WR_DIS_BLOCK_KEY2

[WR_DIS.KEY2] wr_dis of BLOCK_KEY2

WR_DIS_BLOCK_KEY3

[WR_DIS.KEY3] wr_dis of BLOCK_KEY3

WR_DIS_BLOCK_KEY4

[WR_DIS.KEY4] wr_dis of BLOCK_KEY4

WR_DIS_BLOCK_KEY5

[WR_DIS.KEY5] wr_dis of BLOCK_KEY5

WR_DIS_BLOCK_SYS_DATA1

[WR_DIS.SYS_DATA_PART1] wr_dis of BLOCK_SYS_DATA1

WR_DIS_BLOCK_SYS_DATA2

[WR_DIS.SYS_DATA_PART2] wr_dis of BLOCK_SYS_DATA2

WR_DIS_BLOCK_USR_DATA

[WR_DIS.USER_DATA] wr_dis of BLOCK_USR_DATA

WR_DIS_KEY_PURPOSE_0

[WR_DIS.KEY0_PURPOSE] wr_dis of KEY_PURPOSE_0

WR_DIS_KEY_PURPOSE_1

[WR_DIS.KEY1_PURPOSE] wr_dis of KEY_PURPOSE_1

WR_DIS_KEY_PURPOSE_2

[WR_DIS.KEY2_PURPOSE] wr_dis of KEY_PURPOSE_2

WR_DIS_KEY_PURPOSE_3

[WR_DIS.KEY3_PURPOSE] wr_dis of KEY_PURPOSE_3

WR_DIS_KEY_PURPOSE_4

[WR_DIS.KEY4_PURPOSE] wr_dis of KEY_PURPOSE_4

WR_DIS_KEY_PURPOSE_5

[WR_DIS.KEY5_PURPOSE] wr_dis of KEY_PURPOSE_5

WR_DIS_MAC

[WR_DIS.MAC_FACTORY] wr_dis of MAC

WR_DIS_MAC_EXT

[] wr_dis of MAC_EXT

WR_DIS_RD_DIS

[] wr_dis of RD_DIS

WR_DIS_SECURE_BOOT_EN

[] wr_dis of SECURE_BOOT_EN

WR_DIS_SECURE_BOOT_KEY_REVOKE0

[] wr_dis of SECURE_BOOT_KEY_REVOKE0

WR_DIS_SECURE_BOOT_KEY_REVOKE1

[] wr_dis of SECURE_BOOT_KEY_REVOKE1

WR_DIS_SECURE_BOOT_KEY_REVOKE2

[] wr_dis of SECURE_BOOT_KEY_REVOKE2

WR_DIS_SPI_BOOT_CRYPT_CNT

[] wr_dis of SPI_BOOT_CRYPT_CNT

XTS_KEY_LENGTH_256

[] Set this bit to configure flash encryption use xts-128 key; else use xts-256 key