Struct esp_hal::gpio::GpioPin

pub struct GpioPin<MODE, const GPIONUM: u8> { /* private fields */ }

Implementations

impl<MODE, const GPIONUM: u8> GpioPin<MODE, GPIONUM>
where Self: GpioProperties,

pub fn into_floating_input(self) -> GpioPin<Input<Floating>, GPIONUM>

Configures the pin to operate as a floating input pin

pub fn into_pull_up_input(self) -> GpioPin<Input<PullUp>, GPIONUM>

Configures the pin to operate as a pulled up input pin

pub fn into_pull_down_input(self) -> GpioPin<Input<PullDown>, GPIONUM>

Configures the pin to operate as a pulled down input pin

pub fn into_inverted_floating_input( self ) -> GpioPin<InvertedInput<Floating>, GPIONUM>

Configures the pin to operate as an inverted floating input pin. Only suitable to be passed into a peripheral driver.

pub fn into_inverted_pull_up_input( self ) -> GpioPin<InvertedInput<PullUp>, GPIONUM>

Configures the pin to operate as an inverted pulled up input pin. Only suitable to be passed into a peripheral driver.

pub fn into_inverted_pull_down_input( self ) -> GpioPin<InvertedInput<PullDown>, GPIONUM>

Configures the pin to operate as an inverted pulled down input pin. Only suitable to be passed into a peripheral driver.

impl<MODE, const GPIONUM: u8> GpioPin<MODE, GPIONUM>
where Self: GpioProperties, <Self as GpioProperties>::PinType: IsOutputPin,

pub fn into_push_pull_output(self) -> GpioPin<Output<PushPull>, GPIONUM>

Configures the pin to operate as an push pull output pin

pub fn into_open_drain_output(self) -> GpioPin<Output<OpenDrain>, GPIONUM>

Configures the pin to operate as an open drain output pin

pub fn into_inverted_push_pull_output( self ) -> GpioPin<InvertedOutput<PushPull>, GPIONUM>

Configures the pin to operate as an inverted push pull output pin. Only suitable to be passed into an peripheral driver

pub fn into_inverted_open_drain_output( self ) -> GpioPin<InvertedOutput<OpenDrain>, GPIONUM>

Configures the pin to operate as an open drain output pin. Only suitable to be passed into an peripheral driver

pub fn into_alternate_1(self) -> GpioPin<Alternate<AF1>, GPIONUM>

pub fn into_alternate_2(self) -> GpioPin<Alternate<AF2>, GPIONUM>

impl<MODE, const GPIONUM: u8> GpioPin<MODE, GPIONUM>
where Self: GpioProperties, <Self as GpioProperties>::PinType: IsAnalogPin,

pub fn into_analog(self) -> GpioPin<Analog, GPIONUM>

impl<MODE> GpioPin<MODE, 0>

pub fn degrade(self) -> AnyPin<MODE, InputOutputAnalogPinType>

impl<MODE> GpioPin<MODE, 1>

pub fn degrade(self) -> AnyPin<MODE, InputOutputAnalogPinType>

impl<MODE> GpioPin<MODE, 2>

pub fn degrade(self) -> AnyPin<MODE, InputOutputAnalogPinType>

impl<MODE> GpioPin<MODE, 3>

pub fn degrade(self) -> AnyPin<MODE, InputOutputAnalogPinType>

impl<MODE> GpioPin<MODE, 4>

pub fn degrade(self) -> AnyPin<MODE, InputOutputAnalogPinType>

impl<MODE> GpioPin<MODE, 5>

pub fn degrade(self) -> AnyPin<MODE, InputOutputAnalogPinType>

impl<MODE> GpioPin<MODE, 6>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 7>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 8>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 9>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 10>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 11>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 12>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 13>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 14>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 15>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 16>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 17>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 18>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 19>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 20>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 21>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 22>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 23>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 24>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 25>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 26>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>

impl<MODE> GpioPin<MODE, 27>

pub fn degrade(self) -> AnyPin<MODE, InputOutputPinType>