Module esp_hal::peripherals

Peripheral Instances

This module creates singleton instances for each of the various peripherals, and re-exports them to allow users to access and use them in their applications.

Should be noted that that the module also re-exports the Interrupt enum from the PAC, allowing users to handle interrupts associated with these peripherals.

Structs

ADC

AES

ASSIST_DEBUG

AXI_DMA

AXI_ICM

BITSCRAMBLER

CACHE

DMA

DS

ECC

ECDSA

EFUSE

GPIO

GPIO_SD

H264

H264_DMA

HMAC

HP_SYS_CLKRST

I2C0

I2C1

I2S0

I2S1

I2S2

I3C_MST

I3C_MST_MEM

I3C_SLV

INTERRUPT_CORE0

INTERRUPT_CORE1

IO_MUX

ISP

JPEG

LCD_CAM

LEDC

LP_ADC

LP_ANA_PERI

LP_AON_CLKRST

LP_GPIO

LP_HUK

LP_I2C0

LP_I2C_ANA_MST

LP_I2S0

LP_INTR

LP_IO_MUX

LP_PERI

LP_SYS

LP_TIMER

LP_TOUCH

LP_TSENS

LP_UART

LP_WDT

MCPWM0

MCPWM1

MIPI_CSI_BRIDGE

MIPI_CSI_HOST

MIPI_DSI_BRIDGE

MIPI_DSI_HOST

PARL_IO

PAU

PCNT

PMU

PPA

PVT

Peripherals

RMT

RSA

SDHOST

SHA

SOC_ETM

SPI0

SPI1

SPI2

SPI3

SYSTEM

SYSTIMER

TIMG0

TIMG1

TRACE0

TRACE1

TWAI0

TWAI1

TWAI2

UART0

UHCI0

USB_DEVICE

USB_WRAP

Enums

Interrupt

Enumeration of all the interrupts.