Struct esp_hal::peripherals::APB_CTRL

pub struct APB_CTRL { /* private fields */ }

Implementations

impl APB_CTRL

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <APB_CTRL as Deref>::Target = {0x60026000 as *const <esp32s3::APB_CTRL as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <APB_CTRL as Deref>::Target

Return the pointer to the register block