Struct esp_hal::peripherals::ASSIST_DEBUG

pub struct ASSIST_DEBUG { /* private fields */ }

Implementations

impl ASSIST_DEBUG

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <ASSIST_DEBUG as Deref>::Target = {0x600c2000 as *const <esp32h2::ASSIST_DEBUG as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <ASSIST_DEBUG as Deref>::Target

Return the pointer to the register block