Struct esp_hal::peripherals::DMA

pub struct DMA { /* private fields */ }

Implementations

impl DMA

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <DMA as Deref>::Target = {0x50081000 as *const <esp32p4::DMA as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <DMA as Deref>::Target

Return the pointer to the register block