Struct esp_hal::peripherals::EFUSE

pub struct EFUSE { /* private fields */ }

Implementations

impl EFUSE

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <EFUSE as Deref>::Target = {0x3ff5a000 as *const <esp32::EFUSE as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <EFUSE as Deref>::Target

Return the pointer to the register block