Struct esp_hal::peripherals::I2S0

pub struct I2S0 { /* private fields */ }

Implementations

impl I2S0

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <I2S0 as Deref>::Target = {0x6000d000 as *const <esp32h2::I2S0 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <I2S0 as Deref>::Target

Return the pointer to the register block