Struct esp_hal::peripherals::INTERRUPT_CORE0

pub struct INTERRUPT_CORE0 { /* private fields */ }

Implementations

impl INTERRUPT_CORE0

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <INTERRUPT_CORE0 as Deref>::Target = {0x600c2000 as *const <esp32c3::INTERRUPT_CORE0 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <INTERRUPT_CORE0 as Deref>::Target

Return the pointer to the register block