Struct esp_hal::peripherals::LEDC

pub struct LEDC { /* private fields */ }

Implementations

impl LEDC

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <LEDC as Deref>::Target = {0x3ff59000 as *const <esp32::LEDC as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <LEDC as Deref>::Target

Return the pointer to the register block