Struct esp_hal::peripherals::LPWR

pub struct LPWR { /* private fields */ }

Implementations

impl LPWR

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <RTC_CNTL as Deref>::Target = {0x60008000 as *const <esp32s3::RTC_CNTL as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <RTC_CNTL as Deref>::Target

Return the pointer to the register block