Struct esp_hal::peripherals::Peripherals

pub struct Peripherals {

    pub ADC1: ADC1,
    pub ADC2: ADC2,
    pub AES: AES,
    pub APB_CTRL: APB_CTRL,
    pub ASSIST_DEBUG: ASSIST_DEBUG,
    pub BT: BT,
    pub DMA: DMA,
    pub DS: DS,
    pub EFUSE: EFUSE,
    pub EXTMEM: EXTMEM,
    pub GPIO: GPIO,
    pub GPIO_SD: GPIO_SD,
    pub HMAC: HMAC,
    pub I2C0: I2C0,
    pub I2S0: I2S0,
    pub INTERRUPT_CORE0: INTERRUPT_CORE0,
    pub IO_MUX: IO_MUX,
    pub LEDC: LEDC,
    pub LPWR: LPWR,
    pub RMT: RMT,
    pub RNG: RNG,
    pub RSA: RSA,
    pub SENSITIVE: SENSITIVE,
    pub SHA: SHA,
    pub SPI0: SPI0,
    pub SPI1: SPI1,
    pub SPI2: SPI2,
    pub SYSTEM: SYSTEM,
    pub SYSTIMER: SYSTIMER,
    pub TIMG0: TIMG0,
    pub TIMG1: TIMG1,
    pub TWAI0: TWAI0,
    pub UART0: UART0,
    pub UART1: UART1,
    pub UHCI0: UHCI0,
    pub UHCI1: UHCI1,
    pub USB_DEVICE: USB_DEVICE,
    pub WIFI: WIFI,
    pub XTS_AES: XTS_AES,

}

Fields
ADC1: ADC1ADC2: ADC2AES: AESAPB_CTRL: APB_CTRLASSIST_DEBUG: ASSIST_DEBUGBT: BTDMA: DMADS: DSEFUSE: EFUSEEXTMEM: EXTMEMGPIO: GPIOGPIO_SD: GPIO_SDHMAC: HMACI2C0: I2C0I2S0: I2S0INTERRUPT_CORE0: INTERRUPT_CORE0IO_MUX: IO_MUXLEDC: LEDCLPWR: LPWRRMT: RMTRNG: RNGRSA: RSASENSITIVE: SENSITIVESHA: SHASPI0: SPI0SPI1: SPI1SPI2: SPI2SYSTEM: SYSTEMSYSTIMER: SYSTIMERTIMG0: TIMG0TIMG1: TIMG1TWAI0: TWAI0UART0: UART0UART1: UART1UHCI0: UHCI0UHCI1: UHCI1USB_DEVICE: USB_DEVICEWIFI: WIFIXTS_AES: XTS_AES
Implementations

impl Peripherals

pub fn take() -> Self

Returns all the peripherals once

impl Peripherals

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.