Struct esp_hal::peripherals::RMT

pub struct RMT { /* private fields */ }

Implementations

impl RMT

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <RMT as Deref>::Target = {0x500d4000 as *const <esp32p4::RMT as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <RMT as Deref>::Target

Return the pointer to the register block