Struct esp_hal::peripherals::SENSITIVE

pub struct SENSITIVE { /* private fields */ }

Implementations

impl SENSITIVE

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SENSITIVE as Deref>::Target = {0x600c1000 as *const <esp32c2::SENSITIVE as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SENSITIVE as Deref>::Target

Return the pointer to the register block