Struct esp_hal::peripherals::SLCHOST

pub struct SLCHOST { /* private fields */ }

Implementations

impl SLCHOST

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SLCHOST as Deref>::Target = {0x3ff55000 as *const <esp32::SLCHOST as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SLCHOST as Deref>::Target

Return the pointer to the register block