Struct esp_hal::peripherals::SPI3

pub struct SPI3 { /* private fields */ }

Implementations

impl SPI3

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <SPI3 as Deref>::Target = {0x60025000 as *const <esp32s3::SPI3 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <SPI3 as Deref>::Target

Return the pointer to the register block