Struct esp_hal::peripherals::SYSTEM

pub struct SYSTEM { /* private fields */ }

Implementations

impl SYSTEM

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <HP_SYS as Deref>::Target = {0x500e5000 as *const <esp32p4::HP_SYS as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <HP_SYS as Deref>::Target

Return the pointer to the register block