Struct esp_hal::peripherals::SYSTEM

pub struct SYSTEM { /* private fields */ }

Implementations

impl SYSTEM

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <PCR as Deref>::Target = {0x60096000 as *const <esp32h2::PCR as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <PCR as Deref>::Target

Return the pointer to the register block