Struct esp_hal::peripherals::UHCI0

pub struct UHCI0 { /* private fields */ }

Implementations

impl UHCI0

pub unsafe fn steal() -> Self

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub const PTR: *const <UHCI0 as Deref>::Target = {0x3ff54000 as *const <esp32::UHCI0 as core::ops::Deref>::Target}

Pointer to the register block

pub const fn ptr() -> *const <UHCI0 as Deref>::Target

Return the pointer to the register block